Remove outdated test assertion
datetime.time.fromisoformat() supports a trailing 'Z' character as of
Python 3.11. This code aims to support Python 3.8 and newer.

I could change the test assertions based on the Python version, but that
doesn't seem especially useful. It would mainly be testing the built-in
fromisoformat() method rather than this project's code.