Simplify date filtering
It's unclear what a cutoff date of None means. Does it mean today? Or
does it mean a long time ago, which would cause all events to be
included?

Requiring the cutoff to always be a date is clearer. It still defaults
to today, since that's what most users would expect.