Use event URL from scrape source
Generally the pages we scrape from are overviews of upcoming events,
each of which has its own page with more details. Those detail pages are
helpful for getting the time, location, and so forth.

The detail page may also contain an RSVP link. But it's more helpful for
the scraper to output the link to the detail page itself, not the RSVP.
Users who want to know more can visit the detail page, where they can
read about the event and RSVP if they want.