Loosen duplication checks
Sometimes the scraper outputs an event which is very similar to but not
an exact match for a previous event. These create duplicates in the
spreadsheet. I'm loosening the rules a little to try to reduce this.
It's not a prefect solution, and we may make further improvements in the
future.